+++ App.js
@@ -1,20 +1,7 @@
 import React from "react";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { StyleSheet } from "react-native";
 
-import HomeScreen from "./src/screens/HomeScreen";
-
-const Stack = createNativeStackNavigator();
+import Navigation from "./src/navigation/Navigation";
 
 export default function App() {
-  return (
-    <NavigationContainer>
-      <Stack.Navigator initialRouteName="Notes">
-        <Stack.Screen name="Notes" component={HomeScreen} />
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
+  return <Navigation />;
 }
-
-const styles = StyleSheet.create({});
