+++ src/navigation/Navigation.js
@@ -1,0 +1,17 @@
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+import HomeScreen from "../screens/HomeScreen";
+
+const Stack = createNativeStackNavigator();
+
+export default () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="Notes">
+        <Stack.Screen name="Notes" component={HomeScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
