+++ src/screens/HomeScreen.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, Text, StyleSheet } from "react-native";
+import { View, Text } from "react-native";
 
 const HomeScreen = () => {
   return (
@@ -8,7 +8,5 @@ const HomeScreen = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({});
 
 export default HomeScreen;
